pyproject.toml
[project]
name = "goodreads-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.2",
]
